pyproject.toml
[tool.poetry]
name = "book-summary"
version = "0.1.0"
description = ""
authors = ["grp06 <gpickett00@gmail.com>"]
readme = "README.md"
packages = [{include = "book_summary"}]

[tool.poetry.dependencies]
python = "3.11.4"
uvicorn = "^0.22.0"
gunicorn = "^20.1.0"
openai = "^0.27.8"
openai-async = "^0.0.3"
python-dotenv = "^1.0.0"
fastapi = "^0.101.0"
httpx = "^0.24.1"
reportlab = "^4.0.4"

[tool.poetry.dev-dependencies]
black = "^23.7.0" # Move this line here

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
